requirements.txt
absl-py==2.1.0
anykeystore==0.2
attrs==23.2.0
Automat==22.10.0
batchgenerators==0.25
buildtools==1.0.6
causal-conv1d==1.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
constantly==23.10.4
contourpy==1.2.1
crip==1.8.0
cryptacular==1.6.2
cycler==0.12.1
Cython==3.0.11
defusedxml==0.7.1
docopt==0.6.2
einops==0.8.0
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
furl==2.1.3
future==1.0.0
fvcore==0.1.5.post20221221
greenlet==3.0.3
grpcio==1.64.1
h5py==3.13.0
huggingface-hub==0.23.4
hupper==1.12.1
hyperlink==21.0.0
idna==3.4
imageio==2.34.1
incremental==22.10.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
json5==0.12.0
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
mamba-ssm==1.1.3.post1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
PasteDeploy==3.1.0
pbkdf2==1.3
periodictable==1.7.0
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
portalocker==2.10.1
prefetch_generator==1.0.3
protobuf==5.27.1
pydicom==2.4.4
pyparsing==3.1.2
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytorch-wavelets==1.3.0
pytz==2024.1
PyWavelets==1.8.0
PyYAML==6.0.1
redo==2.0.4
regex==2024.5.15
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==2.0.0
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
SimpleITK==2.3.1
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.31
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
tokenizers==0.19.1
torch==2.1.1+cu118
torch_mando==1.0.1
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.4
traceback2==1.4.0
transaction==4.0
transformers==4.41.2
translationstring==1.4
triton==2.1.0
Twisted==24.3.0
typing_extensions==4.9.0
tzdata==2024.1
unittest2==1.1.0
urllib3==1.26.13
velruse==1.1.1
venusian==3.1.0
WebOb==1.8.7
Werkzeug==3.0.3
WTForms==3.1.2
wtforms-recaptcha==0.3.2
yacs==0.1.8
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1
